[refactor]: Merge classes; RacingGame & Memento
